Update routing of frontend and reconfigure styling
Initially the routing of the existing branches for the frontend
only worked with the specific pages or layouts being worked on
in their respective branch. So the paths and routing were changed
so that different layouts and pages can be linked together. The
paths in app-routing.module.ts will be the ones used for any
components that need to route or link to the path (url) of
another component.

Additionally, the content in the styles.scss file were either modified,
removed, or moved to the appropriate style file for a particular
component that it applies to. Bootstrap components were added for
consistency, but are not completed yet so are not fully responsive
(e.g. navbar/sidebar need to collapse/fit in smaller screens).
Styling/bootstrap components and classes that are being implemented
in other branches will be compared to the ones implemented in this
branch/commit and added/removed depending on what works best.