Drop support for SQLite
Because the notification setting's destination field relies on the
arrays feature of PostgreSQL, SQLite will no longer work as a substitute
during development.

Update the instructions on running the site. Show how to run the db with
Docker Compose and how to run the web server both in/out of Docker.