Make the user email field unique but still optional
In order to use the email to log in, it has to be unique. However, it's
nice to still leave it optional for users that don't want to provide
one and don't care about email notifications.

Apparently Django 1.11+ will convert CharField (and subclasses like
EmailField) empty values to NULL if it's nullable, but this only applies
to forms. Since this field is updated via API, a custom field is needed.