Ensure a new device's owner is the current user (#46)
Ignore any value provided for the `owner` field and instead use a
default value, which is the current user. DRF permits unknown fields,
so it does not raise an error for `owner` despite it being hidden. This
is not ideal, but it's not worth the effort to error for unknown fields.