Disable Django debug mode when running via Docker Compose
The image doesn't have Node.js in the final stage, which would be
required for the ng build that runs in watch mode when DEBUG=true.

A major problem with supporting debug mode is that ng build and
collectstatic write new files within the container. This means
the bind mount would have to be made read-write, and that the owner
of the new files would be root rather than the host user. Trying to
create a non-root user within the image proved to be quite annoying.
It still seems possible, so it may happen eventually.