Fix bind mount overwriting static root in image
Only mount the "web" subdirectory to avoid mounting the static root
on the host.

Add a WEBPACK_STATS environment variable which sets the path for the
stats file which django-webpack-loader reads for static file info.
This allows the image to place the file somewhere such that it won't
be overwritten by the host's file when the bind mount is used. The
paths to static files within the file are absolute, so the paths
in the host's stats file wouldn't be compatible.