Make parameters and tests one-to-one
Add a device FK to the parameter model and remove its test field. Add a
one-to-one parameter field to the test model.

Use triggers to ensure the aforementioned device and parameter fields
cannot be updated. This makes it possible for the test history to rely
on its test FK for the parameter. Otherwise, a test history's test may
no longer refer to the parameter that was used at the time of that
specific test run.

An alternative solution would've been to add the parameter field as well
to the test history, but this would complicate the schema. It may also
require a more complex trigger.